Move customizable_navbar to core
