+++ front/app/components/PageForm/messages.ts
@@ -71,4 +71,8 @@ export default defineMessages({
     id: 'app.components.PagesForm.saveSuccess',
     defaultMessage: 'Page successfully saved',
   },
+  navbarItemTitle: {
+    id: 'app.components.PagesForm..navbarItemTitle',
+    defaultMessage: 'Name in navbar',
+  },
 });
