+++ front/app/containers/Admin/pagesAndMenu/containers/PagesMenu/EditNavbarItemForm/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { withRouter, WithRouterProps } from 'utils/cl-router/withRouter';
 
 // components
-import NavbarItemForm, { FormValues } from '../../components/NavbarItemForm';
+import NavbarItemForm, { FormValues } from '../../../components/NavbarItemForm';
 import SectionFormWrapper from 'containers/Admin/pagesAndMenu/components/SectionFormWrapper';
 import { pagesAndMenuBreadcrumb } from 'containers/Admin/pagesAndMenu/breadcrumbs';
 
@@ -11,7 +11,7 @@ import { isNilOrError } from 'utils/helperUtils';
 import { getInitialFormValues, createNavbarItemUpdateData } from './utils';
 
 // services
-import { updateNavbarItem } from '../../../services/navbar';
+import { updateNavbarItem } from 'services/navbar';
 
 // i18n
 import { injectIntl } from 'utils/cl-intl';
@@ -19,7 +19,7 @@ import { WrappedComponentProps } from 'react-intl';
 
 // hooks
 import useAppConfigurationLocales from 'hooks/useAppConfigurationLocales';
-import useNavbarItem from '../../../hooks/useNavbarItem';
+import useNavbarItem from 'hooks/useNavbarItem';
 import useLocalize from 'hooks/useLocalize';
 
 const EditNavbarItemForm = ({
