+++ front/app/containers/Admin/pagesAndMenu/containers/PagesMenu/NavigationSettings/HiddenNavbarItemList/HiddenNavbarItemList.test.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 import clHistory from 'utils/cl-router/history';
 import { fireEvent, render, screen } from 'utils/testUtils/rtl';
 import HiddenNavbarItemList from '.';
-import { addNavbarItem } from '../../../../services/navbar';
+import { addNavbarItem } from 'services/navbar';
 
 jest.mock('services/locale');
 jest.mock('services/appConfiguration');
@@ -14,7 +14,7 @@ let mockNavbarItems = allNavbarItems;
 const mockRemovedDefaultNavbarItems = [];
 
 jest.mock('hooks/useNavbarItems', () => jest.fn(() => mockNavbarItems));
-jest.mock('../../../../hooks/useRemovedDefaultNavbarItems', () =>
+jest.mock('hooks/useRemovedDefaultNavbarItems', () =>
   jest.fn(() => mockRemovedDefaultNavbarItems)
 );
 
@@ -22,7 +22,7 @@ jest.mock('hooks/useCustomPages');
 jest.mock('hooks/useCustomPageSlugById');
 jest.mock('hooks/useLocale');
 
-jest.mock('../../../../services/navbar', () => ({
+jest.mock('services/navbar', () => ({
   addNavbarItem: jest.fn(),
 }));
