+++ front/app/containers/Admin/pagesAndMenu/containers/PagesMenu/NavigationSettings/VisibleNavbarItemList/index.tsx
@@ -2,10 +2,7 @@ import React from 'react';
 
 // services
 import { getNavbarItemSlug, INavbarItem } from 'services/navbar';
-import {
-  removeNavbarItem,
-  reorderNavbarItem,
-} from '../../../../services/navbar';
+import { removeNavbarItem, reorderNavbarItem } from 'services/navbar';
 import { deleteCustomPage } from 'services/customPages';
 
 // components
