+++ front/app/containers/Admin/pagesAndMenu/containers/PagesMenu/NavigationSettings/index.tsx
