+++ front/app/containers/Admin/pagesAndMenu/index.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import messages from './messages';
-import Outlet from 'components/Outlet';
 import HelmetIntl from 'components/HelmetIntl';
+import PagesMenu from './containers/PagesMenu';
 
 const PagesAndMenuIndex = () => {
   return (
     <>
       <HelmetIntl title={messages.pagesMenuMetaTitle} />
       <div id="e2e-pages-menu-container">
-        <Outlet id="app.containers.Admin.pages-menu.index" />
+        <PagesMenu />
       </div>
     </>
   );
