+++ front/app/modules/index.ts
@@ -47,7 +47,6 @@ import eventsWidgetConfiguration from './commercial/events_widget';
 import insightsConfiguration from './commercial/insights';
 import analyticsConfiguration from './commercial/analytics';
 
-import customizableNavbarConfiguration from './free/customizable_navbar';
 import userConfirmationConfiguration from './free/user_confirmation';
 
 import idViennaSamlConfiguration from './commercial/id_vienna_saml';
@@ -216,10 +215,6 @@ export default loadModules([
   {
     configuration: analyticsConfiguration,
     isEnabled: CL_CONFIG['modules']['commercial/analytics'],
-  },
-  {
-    configuration: customizableNavbarConfiguration,
-    isEnabled: CL_CONFIG['modules']['free/customizable_navbar'],
   },
   {
     configuration: userConfirmationConfiguration,
